requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
redis
pycryptodome
pycrypto
requests